MNT: Re-rendered with conda-smithy 2.4.0
Now that this is a noarch Python recipe, it needs to be re-rendered to
take advantage of these features in the feedstock. This cuts down on the
CI resources by using only one build on CircleCI (without a matrix) and
no other CIs. Updates the README to reflect this change as well.